refactor(weather): split base client helpers
Move notification-event fetches, main fetch orchestration, auto-source selection, and source-specific helpers into focused mixins while preserving WeatherClient as the public entry point.

Tested: weather client regression suite 188 passed; uv run ruff check src tests scripts; uv run pyright

Confidence: high
