+++ src/accessiweather/weather_client_base.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-import inspect
 import logging
 import os
 from collections.abc import Sequence
@@ -12,37 +11,26 @@ from datetime import UTC, datetime, timedelta
 import httpx
 
 from . import (
-    weather_client_enrichment as enrichment,
     weather_client_nws as nws_client,
     weather_client_openmeteo as openmeteo_client,
-    weather_client_parsers as parsers,
-    weather_client_trends as trends,
 )
-from .alert_lifecycle import diff_alerts
 from .cache import WeatherDataCache
-from .config.source_priority import SourcePriorityConfig
-from .forecast_confidence import calculate_forecast_confidence
 from .models import (
     AppSettings,
-    AviationData,
     CurrentConditions,
     Forecast,
     HourlyForecast,
     Location,
-    MinutelyPrecipitationForecast,
-    SourceAttribution,
-    SourceData,
     WeatherAlerts,
     WeatherData,
 )
-from .notifications.minutely_precipitation import parse_pirate_weather_minutely_block
-from .pirate_weather_client import PirateWeatherApiError, PirateWeatherClient
+from .pirate_weather_client import PirateWeatherClient
 from .services import EnvironmentalDataClient
-from .units import resolve_auto_unit_system
 from .utils.retry import APITimeoutError, retry_with_backoff
-from .weather_client_alerts import AlertAggregator
-from .weather_client_fusion import DataFusionEngine
-from .weather_client_parallel import ParallelFetchCoordinator
+from .weather_client_auto import WeatherClientAutoMixin
+from .weather_client_fetch import WeatherClientFetchMixin
+from .weather_client_notification import WeatherClientNotificationMixin
+from .weather_client_sources import WeatherClientSourcesMixin
 
 logger = logging.getLogger(__name__)
 
@@ -61,7 +49,12 @@ def _is_unittest_mock(value: object) -> bool:
     return isinstance(value, Mock)
 
 
-class WeatherClient:
+class WeatherClient(
+    WeatherClientSourcesMixin,
+    WeatherClientAutoMixin,
+    WeatherClientFetchMixin,
+    WeatherClientNotificationMixin,
+):
     """Simple async weather API client."""
 
     def __init__(
@@ -493,1028 +486,6 @@ class WeatherClient:
         # Use deduplication for concurrent requests
         return await self._fetch_weather_data_with_dedup(
             location, force_refresh, skip_notifications
-        )
-
-    async def get_notification_event_data(self, location: Location) -> WeatherData:
-        """Fetch only lightweight data needed for alert/discussion/risk notifications."""
-        logger.info("Fetching notification event data for %s", location.name)
-        weather_data = WeatherData(location=location)
-
-        try:
-            if self.data_source in ("auto", "nws") and self._is_us_location(location):
-                # Use discussion-only fetch so a forecast API failure can never
-                # silently suppress AFD update notifications.
-                discussion_task = asyncio.create_task(self._get_nws_discussion_only(location))
-                alerts_task = asyncio.create_task(self._get_nws_alerts(location))
-                discussion_result, alerts_result = await asyncio.gather(
-                    discussion_task, alerts_task, return_exceptions=True
-                )
-                if isinstance(discussion_result, Exception):
-                    raise discussion_result
-                if isinstance(alerts_result, Exception):
-                    raise alerts_result
-                discussion, discussion_issuance_time = discussion_result
-                alerts = alerts_result
-                logger.debug(
-                    "get_notification_event_data: discussion=%s issuance=%s alerts=%s",
-                    "ok" if discussion else "None",
-                    discussion_issuance_time,
-                    len(alerts.alerts) if alerts and alerts.alerts else 0,
-                )
-                weather_data.discussion = discussion
-                weather_data.discussion_issuance_time = discussion_issuance_time
-                weather_data.alerts = alerts or WeatherAlerts(alerts=[])
-            elif self.data_source in ("auto", "pirateweather") and (
-                pirate_weather_client := self._pirate_weather_client_for_location(location)
-            ):
-                current_task = asyncio.create_task(
-                    pirate_weather_client.get_current_conditions(location)
-                )
-                alerts_task = asyncio.create_task(pirate_weather_client.get_alerts(location))
-                current_result, alerts_result = await asyncio.gather(
-                    current_task, alerts_task, return_exceptions=True
-                )
-                if isinstance(current_result, Exception):
-                    raise current_result
-                if isinstance(alerts_result, Exception):
-                    raise alerts_result
-                weather_data.current = current_result
-                weather_data.alerts = alerts_result or WeatherAlerts(alerts=[])
-            else:
-                # openmeteo provides no alerts; also handles misconfigured clients
-                weather_data.alerts = WeatherAlerts(alerts=[])
-
-            # Only fetch PW minutely in the lightweight poll if the user actually
-            # wants precipitation start/stop notifications AND PW is the active source.
-            # This avoids an extra API call every 60s for users who don't use it.
-            if (
-                self.data_source in ("auto", "pirateweather")
-                and self.pirate_weather_client
-                and self.settings
-            ):
-                _want_start = getattr(self.settings, "notify_minutely_precipitation_start", False)
-                _want_stop = getattr(self.settings, "notify_minutely_precipitation_stop", False)
-                _want_likelihood = getattr(self.settings, "notify_precipitation_likelihood", False)
-                if (_want_start or _want_stop or _want_likelihood) and (
-                    self._should_fetch_minutely_precipitation(location)
-                ):
-                    weather_data.minutely_precipitation = await self._get_pirate_weather_minutely(
-                        location
-                    )
-                    self._last_minutely_poll_by_location[self._location_key(location)] = (
-                        self._utcnow()
-                    )
-
-            # In auto mode the full refresh stores AlertAggregator.aggregate_alerts(...)
-            # output in _previous_alerts. The lightweight poll must produce the same
-            # canonical shape or the two paths will alternate and diff_alerts will fire
-            # phantom "new"/"updated" notifications every refresh cycle.
-            if self.data_source == "auto" and weather_data.alerts is not None:
-                weather_data.alerts = AlertAggregator().aggregate_alerts(weather_data.alerts, None)
-
-            loc_key = self._location_key(location)
-            previous_alerts = self._previous_alerts.get(loc_key)
-            if self.data_source in ("auto", "nws"):
-                _cancel_refs = await self._fetch_nws_cancel_references()
-            else:
-                _cancel_refs = set()
-            weather_data.alert_lifecycle_diff = diff_alerts(
-                previous_alerts, weather_data.alerts, confirmed_cancel_ids=_cancel_refs
-            )
-            if weather_data.alerts is not None:
-                self._previous_alerts[loc_key] = weather_data.alerts
-        except Exception as exc:
-            logger.error("Failed to fetch notification event data for %s: %s", location.name, exc)
-            weather_data.alerts = weather_data.alerts or WeatherAlerts(alerts=[])
-
-        return weather_data
-
-    async def _get_pirate_weather_minutely(
-        self, location: Location
-    ) -> MinutelyPrecipitationForecast | None:
-        """Fetch Pirate Weather minutely precipitation when a client is configured."""
-        client = getattr(self, "pirate_weather_client", None)
-        if client is None:
-            return None
-
-        for method_name in ("get_minutely_forecast", "get_forecast"):
-            method = getattr(client, method_name, None)
-            if not callable(method):
-                continue
-            try:
-                result = method(location)
-                if inspect.isawaitable(result):
-                    result = await result
-                if isinstance(result, dict):
-                    units = getattr(client, "units", "si")
-                    if not isinstance(units, str):
-                        units = "si"
-                    return parse_pirate_weather_minutely_block(result, units=units)
-            except TypeError:
-                logger.debug(
-                    "Pirate Weather client method %s has an unsupported signature", method_name
-                )
-            except Exception as exc:
-                logger.debug("Pirate Weather minutely fetch via %s failed: %s", method_name, exc)
-                return None
-
-        return None
-
-    async def _fetch_weather_data_with_dedup(
-        self, location: Location, force_refresh: bool, skip_notifications: bool = False
-    ) -> WeatherData:
-        """
-        Fetch weather data with deduplication tracking.
-
-        This method prevents multiple concurrent requests for the same location.
-        """
-        location_key = self._location_key(location)
-
-        # Check if there's already an in-flight request for this location
-        if not force_refresh and location_key in self._in_flight_requests:
-            # Wait for the existing request to complete
-            logger.debug(f"Request for {location.name} already in flight, waiting for result")
-            return await self._in_flight_requests[location_key]
-
-        # Create a new task for this request
-        if not force_refresh:
-            task = asyncio.create_task(self._do_fetch_weather_data(location, skip_notifications))
-            self._in_flight_requests[location_key] = task
-
-            try:
-                return await task
-            finally:
-                # Clean up completed request from tracking
-                self._in_flight_requests.pop(location_key, None)
-        else:
-            # Force refresh bypasses deduplication
-            return await self._do_fetch_weather_data(location, skip_notifications)
-
-    async def _do_fetch_weather_data(
-        self, location: Location, skip_notifications: bool = False
-    ) -> WeatherData:
-        """
-        Perform the actual weather data fetch.
-
-        This is the core fetch logic separated for deduplication purposes.
-        """
-        # Check if we should use smart auto source (parallel multi-source fetch)
-        if self.data_source == "auto":
-            return await self._fetch_smart_auto_source(location, skip_notifications)
-
-        # Determine which API to use based on data source and location
-        logger.debug("Determining API choice")
-        api_choice = self._determine_api_choice(location)
-        api_name = {
-            "nws": "NWS",
-            "openmeteo": "Open-Meteo",
-            "pirateweather": "Pirate Weather",
-        }.get(api_choice, "NWS")
-        logger.info(f"Using {api_name} API for {location.name} (data_source: {self.data_source})")
-
-        logger.debug("Creating WeatherData object")
-        weather_data = WeatherData(location=location)
-
-        if api_choice == "pirateweather":
-            # Use Pirate Weather API
-            try:
-                pirate_weather_client = self._pirate_weather_client_for_location(location)
-                if not pirate_weather_client:
-                    raise PirateWeatherApiError("Pirate Weather API key not configured")
-
-                # Parallelize API calls for better performance
-                current, forecast, hourly_forecast, alerts = await asyncio.gather(
-                    pirate_weather_client.get_current_conditions(location),
-                    pirate_weather_client.get_forecast(
-                        location,
-                        days=self._get_forecast_days_for_source(location, source="pirateweather"),
-                    ),
-                    pirate_weather_client.get_hourly_forecast(location),
-                    pirate_weather_client.get_alerts(location),
-                )
-
-                weather_data.current = current
-                weather_data.forecast = forecast
-                weather_data.hourly_forecast = hourly_forecast
-                weather_data.discussion = "Forecast discussion not available from Pirate Weather."
-                weather_data.discussion_issuance_time = None
-                weather_data.alerts = alerts
-
-                # Compute alert lifecycle diff
-                _pw_loc_key = self._location_key(location)
-                _pw_prev = self._previous_alerts.get(_pw_loc_key)
-                weather_data.alert_lifecycle_diff = diff_alerts(_pw_prev, alerts)
-                if alerts is not None:
-                    self._previous_alerts[_pw_loc_key] = alerts
-
-                # Set source attribution
-                weather_data.source_attribution = SourceAttribution(
-                    contributing_sources={"pirateweather"},
-                )
-
-                logger.info(f"Successfully fetched Pirate Weather data for {location.name}")
-
-            except PirateWeatherApiError as e:
-                logger.error(f"Pirate Weather API failed for {location.name}: {e}")
-                self._set_empty_weather_data(weather_data)
-
-        elif api_choice == "openmeteo":
-            # Use Open-Meteo API only (user explicitly selected this source)
-            try:
-                current, forecast, hourly_forecast = await self._fetch_openmeteo_data(location)
-
-                weather_data.current = current
-                weather_data.forecast = forecast
-                weather_data.hourly_forecast = hourly_forecast
-                weather_data.discussion = "Forecast discussion not available from Open-Meteo."
-                weather_data.discussion_issuance_time = None
-                weather_data.alerts = WeatherAlerts(alerts=[])  # Open-Meteo doesn't provide alerts
-
-                # Set source attribution for single-source mode
-                weather_data.source_attribution = SourceAttribution(
-                    contributing_sources={"openmeteo"},
-                )
-
-                logger.info(f"Successfully fetched Open-Meteo data for {location.name}")
-
-            except Exception as e:
-                logger.error(f"Open-Meteo API failed for {location.name}: {e}")
-                self._set_empty_weather_data(weather_data)
-        else:
-            # Use NWS API only (user explicitly selected this source)
-            try:
-                use_openmeteo_forecast = self._should_use_openmeteo_for_extended_forecast(
-                    location, source="nws"
-                )
-
-                if use_openmeteo_forecast:
-                    current_task = asyncio.create_task(self._get_nws_current_conditions(location))
-                    forecast_task = asyncio.create_task(self._get_openmeteo_forecast(location))
-                    discussion_task = asyncio.create_task(self._get_nws_discussion_only(location))
-                    alerts_task = asyncio.create_task(self._get_nws_alerts(location))
-                    hourly_task = asyncio.create_task(self._get_nws_hourly_forecast(location))
-
-                    current = await current_task
-                    forecast = await forecast_task
-                    discussion, discussion_issuance_time = await discussion_task
-                    alerts = await alerts_task
-                    hourly_forecast = await hourly_task
-                else:
-                    (
-                        current,
-                        forecast,
-                        discussion,
-                        discussion_issuance_time,
-                        alerts,
-                        hourly_forecast,
-                    ) = await self._fetch_nws_data(location)
-
-                weather_data.current = current
-                weather_data.forecast = forecast
-                weather_data.hourly_forecast = hourly_forecast
-                weather_data.discussion = discussion
-                weather_data.discussion_issuance_time = discussion_issuance_time
-                weather_data.alerts = alerts
-
-                # Compute alert lifecycle diff for NWS single-source path
-                _nws_loc_key = self._location_key(location)
-                _nws_prev = self._previous_alerts.get(_nws_loc_key)
-                _cancel_refs = await self._fetch_nws_cancel_references()
-                weather_data.alert_lifecycle_diff = diff_alerts(
-                    _nws_prev, alerts, confirmed_cancel_ids=_cancel_refs
-                )
-                if alerts is not None:
-                    self._previous_alerts[_nws_loc_key] = alerts
-
-                # Set source attribution for single-source mode
-                weather_data.source_attribution = SourceAttribution(
-                    field_sources={
-                        "forecast_source": "openmeteo" if use_openmeteo_forecast else "nws",
-                        "hourly_source": "nws",
-                    },
-                    contributing_sources={"nws", "openmeteo"}
-                    if use_openmeteo_forecast
-                    else {"nws"},
-                )
-
-                if (current is None or not current.has_data()) and forecast is None:
-                    logger.warning(f"NWS returned empty data for {location.name}")
-                else:
-                    logger.info(f"Successfully fetched NWS data for {location.name}")
-
-            except Exception as e:
-                logger.error(f"NWS API failed for {location.name}: {e}")
-                self._set_empty_weather_data(weather_data)
-
-        if weather_data.has_any_data():
-            # Launch enrichment tasks in parallel
-            enrichment_tasks = self._launch_enrichment_tasks(
-                weather_data, location, skip_notifications
-            )
-            # Await enrichment completion (which includes persisting to cache)
-            await self._await_enrichments(enrichment_tasks, weather_data)
-
-        if not weather_data.has_any_data() and self.offline_cache:
-            cached = self.offline_cache.load(location)
-            if cached:
-                logger.info(f"Using cached weather data for {location.name}")
-                self._remember_weather_data(cached)
-                return cached
-
-        self._remember_weather_data(weather_data)
-        return weather_data
-
-    def _get_auto_mode_api_budget(self) -> str:
-        """Return the validated automatic-mode API budget setting."""
-        budget = getattr(self.settings, "auto_mode_api_budget", "max_coverage")
-        return budget if budget in {"economy", "balanced", "max_coverage"} else "max_coverage"
-
-    @staticmethod
-    def _source_has_core_section(
-        section: CurrentConditions | Forecast | HourlyForecast | None,
-    ) -> bool:
-        """Return True when a current/forecast/hourly section is present and populated."""
-        return bool(section is not None and section.has_data())
-
-    def _source_has_complete_core_data(self, source_data: SourceData | None) -> bool:
-        """Return True when a single-source result includes current, forecast, and hourly data."""
-        if source_data is None:
-            return False
-        return all(
-            self._source_has_core_section(section)
-            for section in (source_data.current, source_data.forecast, source_data.hourly_forecast)
-        )
-
-    @staticmethod
-    def _configured_sources_in_fetch_order(
-        active_sources: Sequence[str],
-        fetchers: dict[str, object],
-        fetched_sources: set[str] | None = None,
-    ) -> list[str]:
-        """Return configured sources that are currently fetchable, preserving user order."""
-        already_fetched = fetched_sources or set()
-        return [
-            source
-            for source in active_sources
-            if source in fetchers and source not in already_fetched
-        ]
-
-    async def _fetch_auto_mode_sources(
-        self,
-        location: Location,
-        coordinator: ParallelFetchCoordinator,
-        fetchers: dict[str, object],
-        requested_sources: Sequence[str],
-    ) -> list[SourceData]:
-        """Fetch a selected subset of automatic-mode sources."""
-        sources_to_fetch = [source for source in requested_sources if source in fetchers]
-        if not sources_to_fetch:
-            return []
-
-        return await coordinator.fetch_all(
-            location=location,
-            fetch_nws=fetchers["nws"]() if "nws" in sources_to_fetch else None,
-            fetch_openmeteo=fetchers["openmeteo"]() if "openmeteo" in sources_to_fetch else None,
-            fetch_pirateweather=(
-                fetchers["pirateweather"]() if "pirateweather" in sources_to_fetch else None
-            ),
-        )
-
-    async def _fetch_smart_auto_source(
-        self, location: Location, skip_notifications: bool = False
-    ) -> WeatherData:
-        """Fetch weather data using staged automatic-source decisions."""
-        logger.info("Using smart auto source for %s", location.name)
-
-        auto_sources_us = getattr(
-            self.settings,
-            "auto_sources_us",
-            ["nws", "openmeteo", "pirateweather"],
-        )
-        auto_sources_international = getattr(
-            self.settings,
-            "auto_sources_international",
-            ["openmeteo", "pirateweather"],
-        )
-
-        config = SourcePriorityConfig(
-            us_default=auto_sources_us,
-            international_default=auto_sources_international,
-        )
-        auto_budget = self._get_auto_mode_api_budget()
-        parallel_timeout = getattr(self.settings, "parallel_fetch_timeout", 5.0)
-        coordinator = ParallelFetchCoordinator(timeout=parallel_timeout)
-        fusion_engine = DataFusionEngine(config)
-        alert_aggregator = AlertAggregator()
-
-        is_us = self._is_us_location(location)
-        active_sources = list(auto_sources_us if is_us else auto_sources_international)
-
-        async def fetch_openmeteo():
-            current, forecast, hourly = await self._fetch_openmeteo_data(location)
-            return (current, forecast, hourly)
-
-        async def fetch_nws():
-            (
-                current,
-                forecast,
-                _discussion,
-                _discussion_time,
-                alerts,
-                hourly,
-            ) = await self._fetch_nws_data(location)
-            return (current, forecast, hourly, alerts)
-
-        async def fetch_pw():
-            pirate_weather_client = self._pirate_weather_client_for_location(location)
-            if not pirate_weather_client:
-                return (None, None, None, None)
-            current = await pirate_weather_client.get_current_conditions(location)
-            forecast = await pirate_weather_client.get_forecast(
-                location,
-                days=self._get_forecast_days_for_source(location, source="pirateweather"),
-            )
-            hourly = await pirate_weather_client.get_hourly_forecast(location)
-            alerts = await pirate_weather_client.get_alerts(location)
-            return (current, forecast, hourly, alerts)
-
-        fetchers: dict[str, object] = {}
-        if "openmeteo" in active_sources:
-            fetchers["openmeteo"] = fetch_openmeteo
-        if is_us and "nws" in active_sources:
-            fetchers["nws"] = fetch_nws
-        if self._pirate_weather_client_for_location(location) and "pirateweather" in active_sources:
-            fetchers["pirateweather"] = fetch_pw
-
-        if auto_budget == "max_coverage":
-            source_results = await self._fetch_auto_mode_sources(
-                location, coordinator, fetchers, active_sources
-            )
-        else:
-            configured_sources = self._configured_sources_in_fetch_order(active_sources, fetchers)
-            primary_source = configured_sources[0] if configured_sources else None
-            primary_sources = (
-                [primary_source] if primary_source and primary_source in fetchers else []
-            )
-            source_results = await self._fetch_auto_mode_sources(
-                location, coordinator, fetchers, primary_sources
-            )
-
-            fetched_sources = {source.source for source in source_results}
-            primary_result = source_results[0] if source_results else None
-            core_complete = self._source_has_complete_core_data(primary_result)
-            needs_us_openmeteo = (
-                is_us
-                and "openmeteo" in fetchers
-                and "openmeteo" not in fetched_sources
-                and (
-                    self._should_use_openmeteo_for_extended_forecast(location, source="auto")
-                    or not core_complete
-                )
-            )
-            needs_intl_alert_source = (
-                not is_us
-                and primary_source == "openmeteo"
-                and "pirateweather" not in fetched_sources
-            )
-            needs_intl_core_fallback = not is_us and not core_complete
-
-            remaining_sources = self._configured_sources_in_fetch_order(
-                active_sources, fetchers, fetched_sources
-            )
-
-            def _pick_secondary_candidate(predicate) -> list[str]:
-                for source in remaining_sources:
-                    if predicate(source):
-                        return [source]
-                return []
-
-            secondary_sources: list[str] = []
-            if needs_us_openmeteo or (auto_budget == "balanced" and is_us and not core_complete):
-                secondary_sources = _pick_secondary_candidate(lambda _source: True)
-            elif needs_intl_alert_source:
-                secondary_sources = _pick_secondary_candidate(
-                    lambda source: source == "pirateweather"
-                )
-            elif needs_intl_core_fallback:
-                secondary_sources = _pick_secondary_candidate(lambda _source: True)
-
-            if secondary_sources:
-                source_results.extend(
-                    await self._fetch_auto_mode_sources(
-                        location, coordinator, fetchers, secondary_sources
-                    )
-                )
-
-        successful_sources = [s for s in source_results if s.success]
-        if not successful_sources:
-            logger.warning("All sources failed for %s, checking cache", location.name)
-            return self._handle_all_sources_failed(location, source_results)
-
-        merged_current, current_attribution = fusion_engine.merge_current_conditions(
-            source_results, location
-        )
-        requested_days = getattr(self.settings, "forecast_duration_days", 7)
-        merged_forecast, forecast_attribution = fusion_engine.merge_forecasts(
-            source_results, location, requested_days=requested_days
-        )
-        merged_hourly, hourly_attribution = fusion_engine.merge_hourly_forecasts(
-            source_results, location
-        )
-
-        _want_start = getattr(self.settings, "notify_minutely_precipitation_start", False)
-        _want_stop = getattr(self.settings, "notify_minutely_precipitation_stop", False)
-        _want_likelihood = getattr(self.settings, "notify_precipitation_likelihood", False)
-        _pw_fetched = any(source.source == "pirateweather" for source in source_results)
-        merged_minutely = None
-        if self.pirate_weather_api_key and (
-            _pw_fetched or _want_start or _want_stop or _want_likelihood
-        ):
-            merged_minutely = await self._get_pirate_weather_minutely(location)
-
-        has_any_data = (
-            self._source_has_core_section(merged_current)
-            or self._source_has_core_section(merged_forecast)
-            or self._source_has_core_section(merged_hourly)
-        )
-        if not has_any_data:
-            logger.warning("All sources returned empty data for %s", location.name)
-            return self._handle_all_sources_failed(location, source_results)
-
-        nws_alerts = None
-        pw_alerts_data = None
-        for source in source_results:
-            if source.source == "nws" and source.alerts:
-                nws_alerts = source.alerts
-            elif source.source == "pirateweather" and source.alerts:
-                pw_alerts_data = source.alerts
-
-        if is_us and nws_alerts is not None:
-            merged_alerts = alert_aggregator.aggregate_alerts(nws_alerts, None)
-        else:
-            merged_alerts = alert_aggregator.aggregate_alerts(nws_alerts, pw_alerts_data)
-
-        _loc_key = self._location_key(location)
-        _prev_alerts = self._previous_alerts.get(_loc_key)
-        _cancel_refs = (
-            await self._fetch_nws_cancel_references() if nws_alerts is not None else set()
-        )
-        _alert_diff = diff_alerts(_prev_alerts, merged_alerts, confirmed_cancel_ids=_cancel_refs)
-        self._previous_alerts[_loc_key] = merged_alerts
-
-        failed_sources = {source.source for source in source_results if not source.success}
-        contributing_sources = (
-            current_attribution.contributing_sources
-            | set(forecast_attribution.values())
-            | set(hourly_attribution.values())
-            | {source.source for source in source_results if source.alerts is not None}
-        )
-        attribution = SourceAttribution(
-            field_sources={
-                **current_attribution.field_sources,
-                **forecast_attribution,
-                **hourly_attribution,
-            },
-            conflicts=current_attribution.conflicts,
-            contributing_sources=contributing_sources,
-            failed_sources=current_attribution.failed_sources | failed_sources,
-        )
-
-        incomplete_sections: set[str] = set()
-        if merged_current is None:
-            incomplete_sections.add("current")
-        if merged_forecast is None:
-            incomplete_sections.add("forecast")
-        if merged_hourly is None:
-            incomplete_sections.add("hourly_forecast")
-
-        if is_us and any(source.source == "nws" for source in source_results):
-            discussion = "Forecast discussion available from NWS for US locations."
-            discussion_issuance_time = None
-        elif is_us:
-            discussion = (
-                "Forecast discussion is unavailable because Automatic mode did not use NWS."
-            )
-            discussion_issuance_time = None
-        else:
-            discussion = "Forecast discussion not available from Open-Meteo."
-            discussion_issuance_time = None
-
-        confidence = calculate_forecast_confidence(source_results)
-        weather_data = WeatherData(
-            location=location,
-            current=merged_current,
-            forecast=merged_forecast,
-            hourly_forecast=merged_hourly,
-            discussion=discussion,
-            discussion_issuance_time=discussion_issuance_time,
-            minutely_precipitation=merged_minutely,
-            alerts=merged_alerts,
-            source_attribution=attribution,
-            incomplete_sections=incomplete_sections,
-            forecast_confidence=confidence,
-        )
-        weather_data.alert_lifecycle_diff = _alert_diff
-
-        if weather_data.has_any_data():
-            enrichment_tasks = self._launch_enrichment_tasks(
-                weather_data, location, skip_notifications
-            )
-            await self._await_enrichments(enrichment_tasks, weather_data)
-
-        if weather_data.has_any_data() and self.offline_cache:
-            self._persist_weather_data(location, weather_data)
-
-        logger.info(
-            "Smart auto source completed for %s: %d sources succeeded",
-            location.name,
-            len(successful_sources),
-        )
-        return weather_data
-
-    def _handle_all_sources_failed(
-        self, location: Location, source_results: list[SourceData]
-    ) -> WeatherData:
-        """
-        Handle the case when all sources fail.
-
-        Returns cached data if available, marked as stale.
-
-        Args:
-            location: The location that failed
-            source_results: List of failed source results
-
-        Returns:
-            Cached WeatherData marked as stale, or empty WeatherData
-
-        """
-        # Try to return cached data
-        if self.offline_cache:
-            cached = self.offline_cache.load(location, allow_stale=True)
-            if cached:
-                cached.stale = True
-                cached.stale_reason = "All weather sources failed"
-                logger.info(f"Returning stale cached data for {location.name}")
-                self._remember_weather_data(cached)
-                return cached
-
-        # No cache available, return empty data with attribution
-        attribution = SourceAttribution()
-        for source in source_results:
-            attribution.failed_sources.add(source.source)
-
-        # Return empty but non-None objects for backward compatibility
-        return WeatherData(
-            location=location,
-            current=CurrentConditions(),
-            forecast=Forecast(periods=[]),
-            hourly_forecast=HourlyForecast(periods=[]),
-            alerts=WeatherAlerts(alerts=[]),
-            source_attribution=attribution,
-            incomplete_sections={"current", "forecast", "hourly_forecast", "alerts"},
-            stale=True,
-            stale_reason="All weather sources failed and no cached data available",
-        )
-
-    def _launch_enrichment_tasks(
-        self, weather_data: WeatherData, location: Location, skip_notifications: bool = False
-    ) -> dict[str, asyncio.Task]:
-        """
-        Launch enrichment tasks that can run concurrently.
-
-        Returns a dictionary of task names to asyncio.Task objects that can be
-        awaited later for progressive updates.
-
-        Args:
-        ----
-            weather_data: The WeatherData object to enrich
-            location: The location for enrichment
-            skip_notifications: If True, skip triggering alert notifications
-
-        Returns:
-        -------
-            Dictionary mapping enrichment names to their tasks
-
-        """
-        tasks = {}
-
-        # Smart enrichments for auto mode
-        if self.data_source == "auto":
-            tasks["sunrise_sunset"] = asyncio.create_task(
-                enrichment.enrich_with_sunrise_sunset(self, weather_data, location)
-            )
-            tasks["nws_discussion"] = asyncio.create_task(
-                enrichment.enrich_with_nws_discussion(self, weather_data, location)
-            )
-
-        # Post-processing enrichments (always run)
-        tasks["environmental"] = asyncio.create_task(
-            enrichment.populate_environmental_metrics(self, weather_data, location)
-        )
-        tasks["aviation"] = asyncio.create_task(
-            enrichment.enrich_with_aviation_data(self, weather_data, location)
-        )
-        tasks["marine"] = asyncio.create_task(
-            enrichment.enrich_with_marine_data(self, weather_data, location)
-        )
-
-        return tasks
-
-    async def _await_enrichments(
-        self, tasks: dict[str, asyncio.Task], weather_data: WeatherData
-    ) -> None:
-        """
-        Await all enrichment tasks and apply final processing.
-
-        Args:
-        ----
-            tasks: Dictionary of enrichment task names to Task objects
-            weather_data: The WeatherData object being enriched
-
-        """
-        # Await all tasks, capturing exceptions
-        results = await asyncio.gather(*tasks.values(), return_exceptions=True)
-
-        # Log any errors from enrichments (non-fatal)
-        for task_name, result in zip(tasks.keys(), results, strict=False):
-            if isinstance(result, Exception):
-                logger.debug(f"Enrichment '{task_name}' failed: {result}")
-
-        # Apply final processing
-        trends.apply_trend_insights(  # pragma: no cover
-            weather_data,
-            self.trend_insights_enabled,
-            self.trend_hours,
-            include_pressure=self.show_pressure_trend,
-        )
-        self._persist_weather_data(weather_data.location, weather_data)
-
-    def _determine_api_choice(self, location: Location) -> str:
-        """Determine which API to use for the given location."""
-        # Validate data source
-        valid_sources = ["auto", "nws", "openmeteo", "pirateweather"]
-        if self.data_source not in valid_sources:
-            logger.warning(f"Invalid data source '{self.data_source}', defaulting to 'auto'")
-            self.data_source = "auto"
-
-        if self.data_source == "pirateweather":
-            if not self.pirate_weather_client:
-                # The lazy key may have cached an empty result from a transient
-                # keyring failure (e.g., right after an MSI update).  Reset and
-                # retry once before giving up.
-                lazy = self._pirate_weather_api_key
-                if hasattr(lazy, "reset"):
-                    lazy.reset()
-                    # Force re-evaluation after reset
-                    self._pirate_weather_client = None
-                if not self.pirate_weather_client:
-                    logger.warning(
-                        "Pirate Weather selected but no API key provided, falling back to auto"
-                    )
-                    return "nws" if self._is_us_location(location) else "openmeteo"
-            return "pirateweather"
-        if self.data_source == "openmeteo":
-            return "openmeteo"
-        if self.data_source == "nws":
-            return "nws"
-        if self.data_source == "auto":
-            # Use NWS for US locations, Open-Meteo for international locations
-            return "nws" if self._is_us_location(location) else "openmeteo"
-        # Fallback for any unexpected cases
-        logger.warning(f"Unexpected data source '{self.data_source}', defaulting to auto")
-        return "nws" if self._is_us_location(location) else "openmeteo"
-
-    @staticmethod
-    def _location_key(location: Location) -> str:
-        """Return a stable string key for a location (used for alert caching)."""
-        return f"{location.latitude:.4f},{location.longitude:.4f}"
-
-    def _resolve_pirate_weather_units(self, location: Location) -> str:
-        """Resolve the Pirate Weather unit bundle for the given location."""
-        preference = (getattr(self.settings, "temperature_unit", "both") or "both").strip().lower()
-        if preference == "auto":
-            unit_system = resolve_auto_unit_system(location)
-            return "uk" if unit_system.value == "uk" else unit_system.value
-        if preference in {"c", "celsius"}:
-            return "ca"
-        return "us"
-
-    def _pirate_weather_client_for_location(self, location: Location) -> PirateWeatherClient | None:
-        """Return a Pirate Weather client configured for the location's effective unit system."""
-        api_key = self.pirate_weather_api_key
-        if not api_key:
-            return None
-
-        units = self._resolve_pirate_weather_units(location)
-        client = self._pirate_weather_client
-        if client is None or client.units != units:
-            client = PirateWeatherClient(api_key, self.user_agent, units=units)
-            self._pirate_weather_client = client
-        return client
-
-    def _is_us_location(self, location: Location) -> bool:
-        """
-        Check if location is within the United States.
-
-        Uses country_code when available for accurate detection. Falls back to
-        coordinate bounds only for clear-cut cases. For ambiguous regions (near US-Canada
-        border), requires country_code to be set - otherwise returns False to avoid
-        misclassifying Canadian cities like Toronto, Montreal, Ottawa as US locations.
-        """
-        country_code = getattr(location, "country_code", None)
-        if country_code:
-            return country_code.upper() == "US"
-
-        lat = location.latitude
-        lon = location.longitude
-
-        # Continental US bounds (approximate)
-        in_continental_bounds = 24.0 <= lat <= 49.0 and -125.0 <= lon <= -66.0
-
-        # Alaska bounds (51-71°N, 130-172°W)
-        in_alaska_bounds = 51.0 <= lat <= 71.5 and -172.0 <= lon <= -130.0
-
-        # Hawaii bounds (18-23°N, 154-161°W)
-        in_hawaii_bounds = 18.0 <= lat <= 23.0 and -161.0 <= lon <= -154.0
-
-        # For clear-cut cases (far from borders), coordinate bounds are reliable
-        # - Hawaii (obvious island)
-        # - Alaska (far northwest, separated from Canada)
-        # - Deep in continental US (away from Canadian/Mexican borders)
-        if in_hawaii_bounds or in_alaska_bounds:
-            return True
-
-        # For continental US, use both lat/lon to identify clearly US vs ambiguous zones
-        if in_continental_bounds:
-            # Canadian border cities (Toronto ~43°N/-79°, Montreal ~45.5°N/-73°, Ottawa ~45.4°N/-75°)
-            # are in the eastern half of the continent (east of ~-95° longitude)
-            # US Pacific Northwest (Seattle ~47.6°N/-122°, Portland ~45.5°N/-122°) is further west
-            # Use lon < -95 as a proxy for "eastern region where Canadian cities cluster"
-            if lat >= 43.0 and lon > -95.0 and lon < -70.0:
-                # Eastern region near Canadian border - require country_code for accurate detection
-                # This catches Canadian cities while allowing US Pacific Northwest
-                logger.debug(
-                    f"Location '{location.name}' ({lat:.2f}, {lon:.2f}) is in eastern "
-                    f"North America near Canadian border but has no country_code. "
-                    f"To ensure correct detection, re-add the location to set country_code via geocoding."
-                )
-                return False
-            return True
-
-        # Outside all US bounds - definitely not US
-        return False
-
-    def _set_empty_weather_data(self, weather_data: WeatherData) -> None:
-        """Set empty weather data when all APIs fail."""
-        weather_data.current = CurrentConditions()
-        weather_data.forecast = Forecast(periods=[])
-        weather_data.hourly_forecast = HourlyForecast(periods=[])
-        weather_data.discussion = "Weather data not available."
-        weather_data.discussion_issuance_time = None
-        weather_data.alerts = WeatherAlerts(alerts=[])
-
-    async def _get_nws_current_conditions(self, location: Location) -> CurrentConditions | None:
-        """Delegate to the NWS client module."""
-        return await nws_client.get_nws_current_conditions(
-            location, self.nws_base_url, self.user_agent, self.timeout, self._get_http_client()
-        )
-
-    async def _get_nws_forecast_and_discussion(
-        self, location: Location
-    ) -> tuple[Forecast | None, str | None, datetime | None]:
-        """Delegate to the NWS client module."""
-        return await nws_client.get_nws_forecast_and_discussion(
-            location, self.nws_base_url, self.user_agent, self.timeout, self._get_http_client()
-        )
-
-    async def _get_nws_discussion_only(
-        self, location: Location
-    ) -> tuple[str | None, datetime | None]:
-        """Fetch only the NWS AFD discussion (no forecast). Used by the notification path."""
-        return await nws_client.get_nws_discussion_only(
-            location, self.nws_base_url, self.user_agent, self.timeout, self._get_http_client()
-        )
-
-    async def _get_nws_alerts(self, location: Location) -> WeatherAlerts | None:
-        """Delegate to the NWS client module."""
-        alert_radius_type = getattr(self.settings, "alert_radius_type", "county")
-        return await nws_client.get_nws_alerts(
-            location,
-            self.nws_base_url,
-            self.user_agent,
-            self.timeout,
-            self._get_http_client(),
-            alert_radius_type=alert_radius_type,
-        )
-
-    async def _get_nws_hourly_forecast(self, location: Location) -> HourlyForecast | None:
-        """Delegate to the NWS client module."""
-        return await nws_client.get_nws_hourly_forecast(
-            location, self.nws_base_url, self.user_agent, self.timeout, self._get_http_client()
-        )
-
-    async def _get_openmeteo_current_conditions(
-        self, location: Location
-    ) -> CurrentConditions | None:
-        """Delegate to the Open-Meteo client module."""
-        return await openmeteo_client.get_openmeteo_current_conditions(
-            location, self.openmeteo_base_url, self.timeout, self._get_http_client()
-        )
-
-    async def _get_openmeteo_forecast(self, location: Location) -> Forecast | None:
-        """Delegate to the Open-Meteo client module."""
-        return await openmeteo_client.get_openmeteo_forecast(
-            location,
-            self.openmeteo_base_url,
-            self.timeout,
-            self._get_http_client(),
-            days=self._get_forecast_days_for_source(location, source="openmeteo"),
-        )
-
-    def _get_forecast_days_for_source(self, location: Location, source: str) -> int:
-        """
-        Return configured forecast days with location/source caps.
-
-        US locations are capped at 7 days to align with NWS limitations.
-        Other sources are capped by their API limits.
-        """
-        configured = getattr(self.settings, "forecast_duration_days", 7)
-        if not isinstance(configured, int):
-            configured = 7
-        configured = max(3, min(configured, 16))
-
-        source_limits = {
-            "openmeteo": 16,
-            "pirateweather": 8,
-            "nws": 7,
-        }
-        return min(configured, source_limits.get(source, 16))
-
-    async def _get_openmeteo_hourly_forecast(self, location: Location) -> HourlyForecast | None:
-        """Delegate to the Open-Meteo client module."""
-        return await openmeteo_client.get_openmeteo_hourly_forecast(
-            location,
-            self.openmeteo_base_url,
-            self.timeout,
-            self._get_http_client(),
-            hours=self._get_hourly_hours_for_pressure_outlook(),
-        )
-
-    def _get_hourly_hours_for_pressure_outlook(self) -> int:
-        """Return enough hourly data for display and pressure trend windows."""
-        hourly_hours = int(getattr(self.settings, "hourly_forecast_hours", 6) or 6)
-        trend_hours = int(getattr(self.settings, "trend_hours", 24) or 24)
-        return max(1, min(max(hourly_hours, trend_hours), 384))
-
-    async def _augment_current_with_openmeteo(
-        self,
-        current: CurrentConditions | None,
-        location: Location,
-    ) -> CurrentConditions | None:
-        """Fill missing current-condition fields using Open-Meteo data when available."""
-        if current is not None and current.has_data():
-            return current
-
-        try:
-            fallback = await self._get_openmeteo_current_conditions(location)
-        except Exception as exc:  # noqa: BLE001
-            logger.debug("Open-Meteo current conditions fallback failed: %s", exc)
-            return current
-
-        if fallback is None:
-            return current
-
-        if current is None:
-            logger.info(
-                "Using Open-Meteo current conditions for %s due to missing NWS data", location.name
-            )
-            # Strip model-derived snow depth — Open-Meteo uses ERA5/GFS which is
-            # notoriously inaccurate for snowpack. Only station observations are reliable.
-            fallback.snow_depth_in = None  # pragma: no cover
-            fallback.snow_depth_cm = None  # pragma: no cover
-            return fallback
-
-        logger.info(
-            "Supplementing NWS current conditions with Open-Meteo data for %s", location.name
-        )
-        return parsers.merge_current_conditions(current, fallback)
-
-    async def get_aviation_weather(
-        self,
-        station_id: str,
-        *,
-        include_sigmets: bool = False,
-        atsu: str | None = None,
-        include_cwas: bool = False,
-        cwsu_id: str | None = None,
-    ) -> AviationData:
-        return await enrichment.get_aviation_weather(
-            self,
-            station_id,
-            include_sigmets=include_sigmets,
-            atsu=atsu,
-            include_cwas=include_cwas,
-            cwsu_id=cwsu_id,
         )
 
     def _persist_weather_data(self, location: Location, weather_data: WeatherData) -> None:
